feat: Initialize repo with microcontrollers
The microcontrollers captured in this very first batch were the ones I
identified as seeming undersupported by the KiCAD standard library,
popular, and reasonably well assumed to be or become available. It
should not be considerd a reference for what microcontrollers are useful
or should be used in your project. I would like to add more controllers
in the future, but for now I am focused on these controllers I have
already inserted into this repository and supporting them